+++ src/components/Footer/Navigation.tsx
@@ -1,11 +1,8 @@
 import Box from "@mui/material/Box";
 
-import constants from "@/app/styles/constants";
 import { footerLinks } from "./constants";
 import NavigationItem from "./NavigationItem";
 import NavigationList from "./NavigationList";
-
-const iconColor = constants.colors.cream;
 
 const Navigation: React.FC = () => {
   return (
@@ -20,7 +17,7 @@ const Navigation: React.FC = () => {
       <NavigationList>
         {footerLinks.map((item) => (
           <NavigationItem key={item.slug} {...item}>
-            <item.icon sx={{ color: iconColor }} />
+            <item.icon sx={{ color: "background.default" }} />
           </NavigationItem>
         ))}
       </NavigationList>